Client: Game timer

+++ common_base/Networking/Messages/net_message_game_state_update.h
@@ -15,8 +15,8 @@ public:
     int current_turn_worm_id;
     int focused_entity_id;
     float wind_speed;
-    float remaining_game_time;
-    float remaining_turn_time;
+    int remaining_game_time;
+    int remaining_turn_time;
     std::vector<WormDto> worms;
     std::vector<ProjectileDto> projectiles;
     std::vector<WorldEventDto> events;
@@ -28,8 +28,8 @@ public:
         int active_entity_id,
         int current_turn_worm_id,
         float wind_speed,
-        float remaining_game_time,
-        float remaining_turn_time
+        int remaining_game_time,
+        int remaining_turn_time
     );
 
     void add(const WormDto& worm);
